+++ libraries/common/src/test/java/androidx/media3/common/MediaItemTest.java
@@ -1067,6 +1067,24 @@ public class MediaItemTest {
     assertThat(restoredMediaItem.localConfiguration).isEqualTo(mediaItem.localConfiguration);
   }
 
+  /** Regression test for internal b/323302460 */
+  @Test
+  public void roundTripViaBundle_withJustNonNullRequestMetadataExtras_restoresAllData() {
+    Bundle extras = new Bundle();
+    extras.putString("key", "value");
+    MediaItem mediaItem =
+        new MediaItem.Builder()
+            .setMediaId("mediaId")
+            .setRequestMetadata(new RequestMetadata.Builder().setExtras(extras).build())
+            .build();
+
+    MediaItem restoredItem = MediaItem.fromBundle(mediaItem.toBundle());
+
+    assertThat(restoredItem).isEqualTo(mediaItem);
+    assertThat(restoredItem.requestMetadata.extras).isNotNull();
+    assertThat(restoredItem.requestMetadata.extras.get("key")).isEqualTo("value");
+  }
+
   @Test
   public void createDefaultMediaItemInstance_checksDefaultValues() {
     MediaItem mediaItem = new MediaItem.Builder().build();